rename: SideBarStyles 스타일 컴포넌트 규칙에 맞도록 수정
SideBarStyle -> SideBarStyles
